refactor: folder name stand-in => customed, gatsby-func => gatsby
