+++ gatsby-theme-oi-wiki/gatsby-ssr.js
@@ -1,6 +1,6 @@
-import WrapRootElement from './src/gatsby-func/WrapRootElement'
-import onRenderBody from './src/gatsby-func/onRenderBody'
-import onPreRenderHTML from './src/gatsby-func/onPreRenderHTML'
+import WrapRootElement from './src/gatsby/WrapRootElement'
+import onRenderBody from './src/gatsby/onRenderBody'
+import onPreRenderHTML from './src/gatsby/onPreRenderHTML'
 import { enableStaticRendering } from 'mobx-react-lite'
 
 enableStaticRendering(true)
