+++ gatsby-theme-oi-wiki/src/components/customed/Details.tsx
