+++ gatsby-theme-oi-wiki/src/components/customed/Img.tsx
