+++ gatsby-theme-oi-wiki/src/components/customed/index.ts
