+++ gatsby-theme-oi-wiki/src/gatsby/WrapRootElement.tsx
