+++ gatsby-theme-oi-wiki/src/templates/Doc.tsx
@@ -7,7 +7,7 @@ import Main from '../components/Main'
 import TocSidebar from '../components/TocSidebar'
 import NavSidebar from '../components/NavSidebar'
 import styled from '@mui/material/styles/styled'
-import getComponents from '../components/stand-in'
+import getComponents from '../components/customed'
 import MDRenderer from '../components/MDRenderer'
 import { HistoryLocation } from '../types/location'
 import Title from '../components/Title'
